Add WalletServerManager
