+++ Client/Frontend/Browser/BrowserViewController/BrowserViewController+WebViewDelegates.swift
@@ -725,7 +725,7 @@ extension BrowserViewController: WKNavigationDelegate {
 
                 }
             }
-            
+
             decisionHandler(.allow)
             return
         }
